identifier/lib/identifier/pull_request.rb: Workaround for GitHub API...
The GitHub API will return at most 250 commits (not a per-call limit).
It is reasonable to assume that we will have Pull Requests larger than
this, as a PR will likely encompass an entire feature branch.

With the goal of "any components that changed in a PR get built", the
workaround included here resolves this issue by building ALL components
if the commit count for a Pull Request is 250.